Create README.md
Requirements
This script was written to run in Python 2.7 and higher -- a dependency of the STIX library. It might work in Python 3, but I have not tested it.

This script requires some python dependencies. You can install these using pip.

Description

This is a script that connects to a TAXII servers discovery service, grabs the STIX document and parses out indicators and observables. Specifically looking for IP's, Websites, Email Addresses, and Hash's.  It outputs messages to the console creates a CEF message which is then sent via syslog to a syslog smartconnector.